net/mlx5: Fix duplicated offload rules issue
If user adds two same filters using skip_sw, the second command will
succeed. To fix it, if the matches are the same, we will compare the
filter handles. If the handles are the same, we consider it as
a replace command and continue. If the handles are not the same,
we consider it as a duplicate add and fail the command.

Change-Id: Ibe42dcf6d225678240ffd19adbe5a0c2a1158ac4
Signed-off-by: Bodong Wang <bodong@mellanox.com>